re:merge api file
